requirements.txt
pydantic
aiohttp
google-adk
google-genai
vertexai
